move traffic sign processing to separate folder
