+++ app/run-4-process.sh
@@ -9,7 +9,6 @@ source ./process-helpers.sh
 echo -e "\e[1m\e[7m PROCESS – START \e[27m\e[21m – Start Time: $(date)\e[0m"
 
 
-run_lua "trafficSigns"
 
 echo "Reminder: The 'bikelanes' table is available only after Postprocessing finished"
 run_lua "roads_bikelanes/roads_bikelanes"
@@ -20,6 +19,9 @@ run_psql "roads_bikelanes/bikelanes/bikelanes"
 # run_lua "legacy_roadClassification/roadClassification"
 # run_lua "legacy_maxspeed/maxspeed"
 run_lua "legacy_lit/lit"
+
+run_lua "trafficSigns/trafficSigns"
+# run_psql "trafficSigns/trafficSigns"
 
 # run_lua "legacy_boundaries/boundaries"
 run_lua "boundaries/boundaries"
